refactor: Removes Use Cases that only redirect calls.
