+++ app/src/main/java/com/oracle/visualize/data/mapper/UserMappers.kt
@@ -2,17 +2,25 @@ package com.oracle.visualize.data.mapper
 
 import com.oracle.visualize.data.datasources.dtos.UserDTO
 import com.oracle.visualize.domain.models.ShareUser
+import com.oracle.visualize.domain.models.ThemePreference
 import com.oracle.visualize.domain.models.User
+import com.oracle.visualize.domain.models.UserType
 
 fun UserDTO.toDomain(): User = User(
     id = id,
-    userType = userType,
     email = email,
     username = username,
     profilePictureURL = profilePictureURL,
-    themePreference = themePreference,
     chartTheme = chartTheme,
-    hiddenVisualizations = hiddenVisualizations
+    hiddenVisualizations = hiddenVisualizations,
+
+    userType = runCatching {
+        UserType.valueOf(userType)
+    }.getOrDefault(UserType.CONSUMER),
+
+    themePreference = runCatching {
+        ThemePreference.valueOf(themePreference)
+    }.getOrDefault(ThemePreference.SYSTEM)
 )
 fun UserDTO.toShareUser(): ShareUser {
     return ShareUser(
