+++ app/src/main/java/com/oracle/visualize/di/AppModule.kt
@@ -21,7 +21,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-/*
+/* EXAMPLE
 @Module
 @InstallIn(SingletonComponent::class)
 object AppModule {
