+++ app/src/main/java/com/oracle/visualize/domain/models/User.kt
@@ -2,17 +2,17 @@ package com.oracle.visualize.domain.models
 
 data class User (
     val id: String,
-    val userType: String,
+    val userType: UserType,
     val email: String,
     val username: String,
     val profilePictureURL: String,
-    val themePreference: String,
+    val themePreference: ThemePreference,
     val chartTheme: String, // Pending: Check variable Type
     val hiddenVisualizations: List<String>
 )
 
 enum class UserType {
-    ADMIN, WRITER, COSTUMER
+    ADMIN, WRITER, CONSUMER
 }
 
 enum class ThemePreference {
