--- app/src/main/java/com/oracle/visualize/domain/usecases/CreateVisualizationUseCase.kt
@@ -1,28 +1,0 @@
-package com.oracle.visualize.domain.usecases
-
-import com.oracle.visualize.domain.models.Team
-import com.oracle.visualize.domain.models.User
-import com.oracle.visualize.domain.repositories.VisualizationRepository
-import kotlinx.serialization.json.JsonObject
-import javax.inject.Inject
-import kotlin.String
-
-class CreateVisualizationUseCase @Inject constructor(
-    private val visualizationRepository: VisualizationRepository
-) {
-    suspend operator fun invoke(
-        authorID: String,
-        title: String,
-        configJSON: String,
-        sharedWithUsers: List<String>,
-        sharedWithTeams: List<String>
-    ){
-        visualizationRepository.createVisualization(
-            authorID,
-            title,
-            configJSON,
-            sharedWithUsers,
-            sharedWithTeams
-        )
-    }
-}
