+++ app/src/main/java/com/oracle/visualize/presentation/screens/shareScreen/ShareAndPostState.kt
@@ -1,4 +1,7 @@
-package com.oracle.visualize.domain.models
+package com.oracle.visualize.presentation.screens.shareScreen
+
+import com.oracle.visualize.domain.models.ShareTeam
+import com.oracle.visualize.domain.models.ShareUser
 
 data class ShareAndPostState(
     val emailQuery: String = "",
