+++ app/src/main/java/com/oracle/visualize/presentation/screens/shareScreen/ShareAndPostViewModel.kt
@@ -3,8 +3,8 @@ package com.oracle.visualize.presentation.screens.shareScreen
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.oracle.visualize.domain.models.ShareUser
-import com.oracle.visualize.domain.usecases.GetUserSuggestionsUseCase
-import com.oracle.visualize.domain.usecases.GetUsersTeamsUseCase
+import com.oracle.visualize.domain.repositories.TeamRepository
+import com.oracle.visualize.domain.repositories.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -16,8 +16,8 @@ import kotlinx.coroutines.launch
 
 @HiltViewModel
 class ShareAndPostViewModel @Inject constructor(
-    private val getUsersTeamsUseCase: GetUsersTeamsUseCase,
-    private val getUserSuggestionsUseCase: GetUserSuggestionsUseCase
+    private val teamRepository: TeamRepository,
+    private val userRepository: UserRepository
 ) : ViewModel(
 
 ) {
@@ -39,7 +39,7 @@ class ShareAndPostViewModel @Inject constructor(
                 .debounce(500)
                 .filter { it.isNotBlank() }
                 .collect { query ->
-                    val results = getUserSuggestionsUseCase.invoke(query.lowercase().trim())
+                    val results = userRepository.getUserSuggestionsByEmail(query.lowercase().trim())
                     updateSuggestions(results)
                 }
         }
@@ -49,8 +49,8 @@ class ShareAndPostViewModel @Inject constructor(
         viewModelScope.launch {
             _uiState.value = ShareUiState.Loading
             try {
-                val myTeams = getUsersTeamsUseCase.getTeamsUserOwns(userID)
-                val teamsImIn = getUsersTeamsUseCase.getTeamsUserIsIn(userID)
+                val myTeams = teamRepository.getTeamsOwnedByUser(userID)
+                val teamsImIn = teamRepository.getTeamsUserIsIn(userID)
                 val suggestedUsers = emptyList<ShareUser>()
